requirements.txt
asgiref==3.9.1
Django==5.2.6
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
